Move Guid.h to interface
